// frontend/src/context/types/chat.ts
export enum CreateChatType {
	CREATE_CHAT_REQUEST = 'CREATE_CHAT_REQUEST',
	CREATE_CHAT_SUCCESS = 'CREATE_CHAT_SUCCESS',
	CREATE_CHAT_FAIL = 'CREATE_CHAT_FAIL',
}

export enum GetUserChatsType {
	GET_USER_CHATS_REQUEST = 'GET_USER_CHATS_REQUEST',
	GET_USER_CHATS_SUCCESS = 'GET_USER_CHATS_SUCCESS',
	GET_USER_CHATS_FAIL = 'GET_USER_CHATS_FAIL',
}

export enum GetChatType {
	GET_CHAT_REQUEST = 'GET_CHAT_REQUEST',
	GET_CHAT_SUCCESS = 'GET_CHAT_SUCCESS',
	GET_CHAT_FAIL = 'GET_CHAT_FAIL',
}

export enum UpdateChatType {
	UPDATE_CHAT_REQUEST = 'UPDATE_CHAT_REQUEST',
	UPDATE_CHAT_SUCCESS = 'UPDATE_CHAT_SUCCESS',
	UPDATE_CHAT_FAIL = 'UPDATE_CHAT_FAIL',
}

export enum GetChatMessagesType {
	GET_CHAT_MESSAGES_REQUEST = 'GET_CHAT_MESSAGES_REQUEST',
	GET_CHAT_MESSAGES_SUCCESS = 'GET_CHAT_MESSAGES_SUCCESS',
	GET_CHAT_MESSAGES_FAIL = 'GET_CHAT_MESSAGES_FAIL',
}

export enum MarkChatAsReadType {
	MARK_CHAT_AS_READ_REQUEST = 'MARK_CHAT_AS_READ_REQUEST',
	MARK_CHAT_AS_READ_SUCCESS = 'MARK_CHAT_AS_READ_SUCCESS',
	MARK_CHAT_AS_READ_FAIL = 'MARK_CHAT_AS_READ_FAIL',
}

export enum CreateMessageType {
	CREATE_MESSAGE_REQUEST = 'CREATE_MESSAGE_REQUEST',
	CREATE_MESSAGE_SUCCESS = 'CREATE_MESSAGE_SUCCESS',
	CREATE_MESSAGE_FAIL = 'CREATE_MESSAGE_FAIL',
}